backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
asyncpg
pydantic-settings
pika
psycopg2-binary